==> nbs/core/01_clean.ipynb <==
# 01 Clean Directory

> Removes all files from `const.pre_output_path`

#|default_exp core.01_clear_dir

#|hide
import nbdev; nbdev.nbdev_export()

#|hide
from nbdev.showdoc import show_doc

#|export
import pubcrawler as proj
from pubcrawler import const, log, utils, tools
import adu_proj.utils as adutils

#|export
import os
import shutil

#|export
def delete_all_files_from_dir(target_dir):
    # Ensure the directory exists
    if not os.path.exists(target_dir):
        print("The specified directory does not exist.")
        return
    # Loop through all the files and subdirectories in the target directory
    for filename in os.listdir(target_dir):
        file_path = os.path.join(target_dir, filename)
        try:
            # If it's a directory, use shutil.rmtree to delete it and its contents
            if os.path.isdir(file_path):
                shutil.rmtree(file_path)
            # If it's a file, use os.unlink (or os.remove) to delete it
            else:
                os.unlink(file_path)
        except Exception as e:
            print("Failed to delete %s. Reason: %s" % (file_path, e))

#|export
delete_all_files_from_dir(const.pre_output_path)
