==> nbs/core/06_archive.ipynb <==
# 06 Archive

> Zip all files.

#|default_exp core.06_archive

#|hide
import nbdev; nbdev.nbdev_export()

#|hide
from nbdev.showdoc import show_doc

#|export
import pubcrawler as proj
from pubcrawler import const, log, utils, tools
import adu_proj.utils as adutils

#|export
import shutil
import boto3
import os

Export all data to zip:

#|export
if shutil.make_archive(f'{const.data_path}/{const.directory_name}', 'zip', f'{const.pre_output_path}'):
    print(f"All files saved to {const.data_path}/{const.directory_name}")

#|export
def upload_zip_to_s3(file_path, bucket_name, s3_file_name):
    # Create an S3 client
    s3_client = boto3.client('s3')

    # Open the file and upload to S3
    with open(file_path, 'rb') as file:
        s3_client.upload_fileobj(file, bucket_name, s3_file_name)

#|export
if const.s3_bucket:
    upload_zip_to_s3(f"{const.data_path}/{const.directory_name}.zip", const.s3_bucket, f"{const.directory_name}.zip")
